refactor: 스터디 탭 섹션 및 강의 목록 컴포넌트 리팩토링 #112
- 관리자 앱의 스터디 탭 섹션 컴포넌트 재구성, 전용 탭(LectureListTab, StudyAnnouncementTab, StudyMemberTab) 및 다이얼로그 래퍼로 로직 분리
- 개선된 관리자/학생 상호작용 및 모바일/데스크톱 지원을 위한 UI 패키지의 강의 목록 및 아이템 컴포넌트 업데이트
- 레거시 tab-section 및 lecture-list 파일 제거, 일관성을 위한 props 및 콜백 시그니처 조정
- 새로운 다이얼로그 래퍼 API에 맞춘 인트라 앱 업데이트
