+++ apps/admin/src/features/study/components/index.ts
@@ -1,5 +1,5 @@
 // Study API exports
 export { CreateStudyForm } from './create-study-form-section/create-study-form';
 export { StudyTable } from './study-table/study-table';
-export { TabSection } from './tab-section/tab-section';
-export { UpdateStatusDialog } from './tab-section/update-status-dialog';
+export { TabSection } from './tab-section/TabSection';
+export { UpdateStatusDialog } from './tab-section/dialog/UpdateStatusDialog';
