+++ packages/ui/src/components/study/LectureList/LectureList.tsx
@@ -1,4 +1,4 @@
-import { cn, LectureListItem } from '@hiarc-platform/ui';
+import { cn, Label, LectureListItem } from '@hiarc-platform/ui';
 import { Lecture } from '@hiarc-platform/shared';
 
 interface LectureListProps {
@@ -13,9 +13,9 @@ interface LectureListProps {
   onTitleClick?(lecture: Lecture): void;
   onAttendanceCheckClick?(lecture: Lecture): void;
   onDoAssignmentClick?(lecture: Lecture): void;
-  onCreateAttendanceClick?(lecture: Lecture, onSuccess: () => void): void;
+  onCreateAttendanceClick?(lecture: Lecture): void;
   onShowAttendanceClick?(lecture: Lecture): void;
-  onCreateAssignmentClick?(lecture: Lecture, onSuccess?: () => void): void;
+  onCreateAssignmentClick?(lecture: Lecture): void;
   onShowAssignmentClick?(lecture: Lecture): void;
   onEditClick?(lecture: Lecture): void;
   onDeleteClick?(lecture: Lecture): void;
@@ -37,7 +37,7 @@ export function LectureList({
   onDeleteClick,
 }: LectureListProps): React.ReactElement {
   return (
-    <div className={cn('flex w-full flex-col gap-2', className)}>
+    <section className={cn('flex w-full flex-col gap-2', className)}>
       {lectureList && lectureList.length > 0 ? (
         lectureList.map((lecture) => (
           <LectureListItem
@@ -45,38 +45,20 @@ export function LectureList({
             isStudent={isStudent}
             key={lecture.round}
             lecture={lecture}
-            onTitleClick={() => {
-              onTitleClick?.(lecture);
-            }}
-            onAttendanceCheckClick={() => {
-              onAttendanceCheckClick?.(lecture);
-            }}
-            onDoAssignmentClick={() => {
-              onDoAssignmentClick?.(lecture);
-            }}
-            onCreateAttendanceClick={(onSuccess) => {
-              onCreateAttendanceClick?.(lecture, onSuccess);
-            }}
-            onShowAttendanceClick={() => {
-              onShowAttendanceClick?.(lecture);
-            }}
-            onCreateAssignmentClick={(onSuccess) => {
-              onCreateAssignmentClick?.(lecture, onSuccess);
-            }}
-            onShowAssignmentClick={() => {
-              onShowAssignmentClick?.(lecture);
-            }}
-            onEditClick={() => {
-              onEditClick?.(lecture);
-            }}
-            onDeleteClick={() => {
-              onDeleteClick?.(lecture);
-            }}
+            onTitleClick={() => onTitleClick?.(lecture)}
+            onAttendanceCheckClick={() => onAttendanceCheckClick?.(lecture)}
+            onDoAssignmentClick={() => onDoAssignmentClick?.(lecture)}
+            onCreateAttendanceClick={() => onCreateAttendanceClick?.(lecture)}
+            onShowAttendanceClick={() => onShowAttendanceClick?.(lecture)}
+            onCreateAssignmentClick={() => onCreateAssignmentClick?.(lecture)}
+            onShowAssignmentClick={() => onShowAssignmentClick?.(lecture)}
+            onEditClick={() => onEditClick?.(lecture)}
+            onDeleteClick={() => onDeleteClick?.(lecture)}
           />
         ))
       ) : (
-        <div className="p-4 text-center text-gray-500">커리큘럼이 없습니다.</div>
+        <Label className="p-4 text-center text-gray-500">커리큘럼이 없습니다.</Label>
       )}
-    </div>
+    </section>
   );
 }
