+++ packages/ui/src/components/study/index.ts
@@ -1,1 +1,2 @@
 export { StudyAnnouncementTable } from './StudyAnnouncementTable/StudyAnnouncementTable';
+export { StudentList } from './StudentList';
