+++ packages/ui/src/index.ts
@@ -28,7 +28,6 @@ export * from './components/label/title';
 export * from './components/layout/content-section';
 export * from './components/layout/page-layout';
 export * from './components/study';
-export * from './components/study/lecture-list-item';
 export * from './components/study/study-unassigned-group';
 export * from './components/study/add-group-dialog/add-group-dialog';
 export * from './components/study/edit-group-dialog/edit-group-dialog';
@@ -59,6 +58,8 @@ export * from './components/study/mobile-study-button';
 export * from './components/study/study-group-list-item';
 export * from './components/study/study-group-list';
 export * from './components/study/StudyAnnouncementTable/StudyAnnouncementTable';
+export * from './components/study/LectureList/LectureList';
+export * from './components/study/LectureListItem/LectureListItem';
 export * from './components/select/checkbox-list';
 export * from './components/select/labeled-checkbox-list';
 export * from './constants/assets';
